% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finn.R
\name{finn}
\alias{finn}
\title{Title}
\usage{
finn(expose, samplesize = NULL, pval = 5e-08)
}
\arguments{
\item{expose}{the variant you need to read from fread}

\item{samplesize}{if you want to calculate the Fval}

\item{pval}{pval}
}
\value{
expose_new
}
\description{
Title
}
