% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fvalue.R
\name{Fval}
\alias{Fval}
\title{Calculate F value}
\usage{
Fval(
  data,
  eaf = "eaf.exposure",
  beta = "beta.exposure",
  se = "se.exposure",
  N = "samplesize.exposure",
  path = "",
  filename = "Fval.csv"
)
}
\arguments{
\item{data}{expose data,it's better to use TwosampleMR read_exposed_data's data}

\item{eaf}{frequency col}

\item{beta}{effective col}

\item{se}{se col}

\item{N}{samplesize col}

\item{path}{the way you save the expose data}

\item{filename}{the name you save the expose data}
}
\value{
the total F value, and a file of the expose data which includ the F value respectively
and a select column which based on the F value more than 10.
}
\description{
Calculate F value
}
